Refactor Staking Provider (#1618)
* refactor: move user staking data into its own composable which gets consumed by the parent staking provider

* cleanup: remove file that should be in apr PR

* fix: lint
