+++ src/components/contextual/pages/pool/MyPoolBalancesCard.vue
@@ -38,7 +38,9 @@ const { isWalletReady } = useWeb3();
 const { isStableLikePool, isStablePhantomPool, isMigratablePool } = usePool(
   toRef(props, 'pool')
 );
-const { stakedSharesForProvidedPool } = useStaking();
+const {
+  userData: { stakedSharesForProvidedPool }
+} = useStaking();
 const router = useRouter();
 
 /**
