+++ src/components/contextual/pages/pools/UnstakedPoolsTable.vue
@@ -22,10 +22,12 @@ const stakePool = ref<FullPool | undefined>();
 
 /** COMPOSABLES */
 const {
-  userGaugeShares,
-  userLiquidityGauges,
-  stakedPools,
-  isLoadingStakingData,
+  userData: {
+    userGaugeShares,
+    userLiquidityGauges,
+    stakedPools,
+    isLoadingUserStakingData
+  },
   setPoolAddress
 } = useStaking();
 
@@ -130,7 +132,9 @@ function handleModalClose() {
     <h5 v-if="!isL2">{{ $t('staking.unstakedPools') }}</h5>
     <PoolsTable
       :key="poolsToRender"
-      :isLoading="isLoadingStakingData || isLoadingUserPools || isUserPoolsIdle"
+      :isLoading="
+        isLoadingUserStakingData || isLoadingUserPools || isUserPoolsIdle
+      "
       :data="poolsToRender"
       :noPoolsLabel="$t('noInvestments')"
       :hiddenColumns="hiddenColumns"
