+++ src/composables/staking/useStaking.ts
@@ -1,6 +1,6 @@
 import { inject } from 'vue';
 
-import { StakingProviderSymbol } from '@/providers/local/staking.provider';
+import { StakingProviderSymbol } from '@/providers/local/staking/staking.provider';
 
 export default function useStaking() {
   const providedData = inject(StakingProviderSymbol);
