+++ src/pages/_layouts/DefaultLayout.vue
@@ -5,7 +5,7 @@ import { useRoute } from 'vue-router';
 import AppHero from '@/components/heros/AppHero.vue';
 import AppNav from '@/components/navs/AppNav/AppNav.vue';
 import useBreakpoints from '@/composables/useBreakpoints';
-import StakingProvider from '@/providers/local/staking.provider';
+import StakingProvider from '@/providers/local/staking/staking.provider';
 
 /**
  * COMPOSABLES
